# microservices/03-product-catalog-service/openapi.json
...
openapi: 3.0.3
info:
  title: Product Catalog Service API
  description: スキー用品ショップ 商品カタログサービス API
  contact:
    email: dev@ski-shop.com
  version: 1.0.0
servers:
- url: "{\"servers\": [{\"url\": \"https://api.ski-shop.com\""
- url: "\"description\": \"Production\"}]}"
...
paths:
...
  /api/v1/categories/level/{level}:
    get:
      tags:
      - Categories
      summary: レベル別カテゴリ取得
      description: 指定されたレベルのカテゴリ一覧を商品数と共に取得します
      parameters:
      - name: level
        in: path
        required: true
        schema:
          format: int32
          type: integer
      responses:
        "200":
          description: 指定レベルのカテゴリ一覧
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryWithProductCountResponse'
...
  /api/v1/categories/path:
    get:
      tags:
      - Categories
      summary: パスでカテゴリ取得
      description: 指定されたパスのカテゴリ詳細を取得します
      parameters:
      - name: path
        in: query
        schema:
          type: string
      responses:
        "200":
          description: カテゴリ詳細
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        "404":
          description: カテゴリが見つかりません
        "400":
          description: パスパラメータが必須です
...
  /api/v1/categories/{categoryId}:
    get:
      tags:
      - Categories
      summary: カテゴリ詳細取得
      description: 指定されたIDのカテゴリ詳細を取得します
      parameters:
      - name: categoryId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      responses:
        "200":
          description: カテゴリ詳細
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        "404":
          description: カテゴリが見つかりません
  /api/v1/categories/{categoryId}/children:
    get:
      tags:
      - Categories
      summary: 子カテゴリ一覧取得
      description: 指定されたカテゴリの子カテゴリ一覧を商品数と共に取得します
      parameters:
      - name: categoryId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      responses:
        "200":
          description: 子カテゴリ一覧
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryWithProductCountResponse'
  /api/v1/categories/{categoryId}/products:
    get:
      tags:
      - Categories
      summary: カテゴリの商品一覧取得
      description: 指定されたカテゴリに属する商品一覧を取得します
      parameters:
      - name: categoryId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      - name: limit
        in: query
        schema:
          format: int32
          default: 10
          type: integer
      responses:
        "200":
          description: カテゴリと商品一覧
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryWithProductsResponse'
        "404":
          description: カテゴリが見つかりません
  /api/v1/categories/{categoryId}/subcategories:
    get:
      tags:
      - Categories
      summary: サブカテゴリ一覧取得
      description: 指定されたカテゴリのサブカテゴリ一覧を商品数と共に取得します
      parameters:
      - name: categoryId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      responses:
        "200":
          description: サブカテゴリ一覧
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryWithProductCountResponse'
  /api/v1/categories/{categoryId}/subcategories/products:
    get:
      tags:
      - Categories
      summary: サブカテゴリ毎の商品一覧取得
      description: 指定されたカテゴリのサブカテゴリ毎に商品一覧を取得します
      parameters:
      - name: categoryId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      - name: limit
        in: query
        schema:
          format: int32
          default: 5
          type: integer
      responses:
        "200":
          description: サブカテゴリ毎の商品一覧
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryWithProductsResponse'
        "404":
          description: カテゴリが見つかりません
...
  /api/v1/products/category/{categoryId}:
    get:
      tags:
      - Products
      summary: カテゴリ別商品一覧
      description: 指定されたカテゴリの商品一覧を取得します
      parameters:
      - name: categoryId
        in: path
        description: カテゴリID
        required: true
        schema:
          $ref: '#/components/schemas/UUID'
      responses:
        "200":
          description: カテゴリ別商品一覧
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductSummaryResponse'
...
components:
  schemas:
    BrandSummaryResponse:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
        logoUrl:
          type: string
        country:
          type: string
    CategoryResponse:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
        description:
          type: string
        path:
          type: string
        level:
          format: int32
          type: integer
        sortOrder:
          format: int32
          type: integer
        imageUrl:
          type: string
        isActive:
          type: boolean
        productCount:
          format: int64
          type: integer
        parent:
          $ref: '#/components/schemas/CategorySummaryResponse'
        children:
          type: array
          items:
            $ref: '#/components/schemas/CategorySummaryResponse'
        createdAt:
          $ref: '#/components/schemas/LocalDateTime'
        updatedAt:
          $ref: '#/components/schemas/LocalDateTime'
        root:
          type: boolean
        subCategory:
          type: boolean
        subCategoryCount:
          format: int32
          type: integer
        active:
          type: boolean
    CategorySummaryResponse:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
        path:
          type: string
    CategoryWithProductCountResponse:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
        description:
          type: string
        path:
          type: string
        level:
          format: int32
          type: integer
        sortOrder:
          format: int32
          type: integer
        imageUrl:
          type: string
        isActive:
          type: boolean
        productCount:
          format: int64
          type: integer
        active:
          type: boolean
    CategoryWithProductsResponse:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
        description:
          type: string
        path:
          type: string
        level:
          format: int32
          type: integer
        sortOrder:
          format: int32
          type: integer
        imageUrl:
          type: string
        active:
          type: boolean
        productCount:
          format: int64
          type: integer
        products:
          type: array
          items:
            $ref: '#/components/schemas/ProductSummaryResponse'
...
    LocalDateTime:
      format: date-time
      type: string
      example: 2022-03-10T12:15:50
...
    ProductSummaryResponse:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        sku:
          type: string
        name:
          type: string
        shortDescription:
          type: string
        category:
          $ref: '#/components/schemas/CategorySummaryResponse'
        brand:
          $ref: '#/components/schemas/BrandSummaryResponse'
        currentPrice:
          type: number
        basePrice:
          type: number
        isOnSale:
          type: boolean
        discountPercentage:
          format: int32
          type: integer
        primaryImageUrl:
          type: string
        inStock:
          type: boolean
        featured:
          type: boolean
        rating:
          format: double
          type: number
        reviewCount:
          format: int32
          type: integer
        tags:
          uniqueItems: true
          type: array
          items:
            type: string
        createdAt:
          $ref: '#/components/schemas/LocalDateTime'
        onSale:
          type: boolean
...
    UUID:
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
      type: string
  securitySchemes:
    SecurityScheme:
      type: http
      description: Authentication
      scheme: bearer
      bearerFormat: JWT